backend/pyproject.toml
[project]
name = "llms-txt-generator"
version = "0.1.0"
description = "llms.txt Generator API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "sse-starlette>=2.0.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
